// src/findjobnu-api/models/JobAgentFrequency.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FindjobnuService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const JobAgentFrequency = {
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3
} as const;
export type JobAgentFrequency = typeof JobAgentFrequency[keyof typeof JobAgentFrequency];


export function instanceOfJobAgentFrequency(value: any): boolean {
    for (const key in JobAgentFrequency) {
        if (Object.prototype.hasOwnProperty.call(JobAgentFrequency, key)) {
            if (JobAgentFrequency[key as keyof typeof JobAgentFrequency] === value) {
                return true;
            }
        }
    }
    return false;
}

export function JobAgentFrequencyFromJSON(json: any): JobAgentFrequency {
    return JobAgentFrequencyFromJSONTyped(json, false);
}

export function JobAgentFrequencyFromJSONTyped(json: any, ignoreDiscriminator: boolean): JobAgentFrequency {
    return json as JobAgentFrequency;
}

export function JobAgentFrequencyToJSON(value?: JobAgentFrequency | null): any {
    return value as any;
}

export function JobAgentFrequencyToJSONTyped(value: any, ignoreDiscriminator: boolean): JobAgentFrequency {
    return value as JobAgentFrequency;
}